--- testproject/Assets/Tests/Runtime/MultiInstanceHelpers.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d96ba9bd07cfb7c4c94a2a81ea590f1e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
